# pied_piper_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pied Piper",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/post/create/{type}": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "create post",
                "description": "User create post here",
                "operationId": "create post",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "normal or job or profilePicture or coverPicture",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "text": {
                                        "type": "text"
                                    },
                                    "images[]": {
                                        "type": "file"
                                    },
                                    "video": {
                                        "type": "file"
                                    },
                                    "voice_record": {
                                        "type": "file"
                                    },
                                    "background": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "post created successfully!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/post/update/{post_id}": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "update post",
                "description": "User update post here",
                "operationId": "update post",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "id of the post",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "text": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "post Updated successfully!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/post/delete/{post_id}": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "delete post",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "id of post you want to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "post deleted"
                    },
                    "401": {
                        "description": "unauthorized."
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/post/friendsPosts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "get all posts",
                "operationId": "friendsPosts",
                "responses": {
                    "200": {
                        "description": "all posts"
                    },
                    "401": {
                        "description": "unauthorized."
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}